organize ModulePlayerChanges
