+++ src/main/java/com/voltskiya/misc/players/attributes/PlayerAttributeListener.java
@@ -1,7 +1,7 @@
-package com.voltskiya.misc.fix;
+package com.voltskiya.misc.players.attributes;
 
 import com.voltskiya.misc.VoltskiyaPlugin;
-import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
@@ -15,6 +15,8 @@ public class PlayerAttributeListener implements Listener {
 
     @EventHandler
     public void playerJoin(PlayerJoinEvent event) {
-        PlayerAttributesConfig.get().updatePlayerAttributes(event.getPlayer());
+        Player player = event.getPlayer();
+        player.setShieldBlockingDelay(0);
+        PlayerAttributesConfig.get().updatePlayerAttributes(player);
     }
 }
